@@ -11,7 +11,7 @@
     "dashboarding"
   ],
   "private": false,
-  "version": "5.0.0-alpha5",
+  "version": "5.0.0-alpha6",
   "build": {
     "number": 8467,
     "sha": "6cb7fec4e154faa0a4a3fee4b33dfef91b9870d9"
@@ -68,7 +68,7 @@
     "url": "https://github.com/elastic/kibana.git"
   },
   "dependencies": {
-    "@bigfunger/decompress-zip": "0.2.0-stripfix2",
+    "@bigfunger/decompress-zip": "0.2.0-stripfix3",
     "@bigfunger/jsondiffpatch": "0.1.38-webpack",
     "@elastic/datemath": "2.3.0",
     "@spalger/filesaver": "1.1.2",
@@ -84,7 +84,7 @@
     "angular-route": "1.4.7",
     "angular-sanitize": "1.5.7",
     "ansicolors": "0.3.2",
-    "autoprefixer": "5.1.1",
+    "autoprefixer": "6.3.7",
     "autoprefixer-loader": "2.0.0",
     "babel": "5.8.38",
     "babel-core": "5.8.38",
@@ -100,8 +100,8 @@
     "csv-parse": "1.1.0",
     "d3": "3.5.6",
     "dragula": "3.7.0",
-    "elasticsearch": "12.0.0-rc4",
-    "elasticsearch-browser": "12.0.0-rc4",
+    "elasticsearch": "12.0.0-rc5",
+    "elasticsearch-browser": "12.0.0-rc5",
     "even-better": "7.0.2",
     "expiry-js": "0.1.7",
     "exports-loader": "0.6.2",
@@ -125,8 +125,8 @@
     "json-stringify-safe": "5.0.1",
     "jstimezonedetect": "1.0.5",
     "leaflet": "0.7.5",
-    "less": "2.5.1",
-    "less-loader": "2.2.0",
+    "less": "2.7.0",
+    "less-loader": "2.2.3",
     "lodash": "3.10.1",
     "marked": "0.3.5",
     "minimatch": "2.0.10",
@@ -159,23 +159,23 @@
     "angular-mocks": "1.4.7",
     "auto-release-sinon": "1.0.3",
     "babel-eslint": "4.1.8",
-    "chokidar": "1.4.3",
-    "chromedriver": "2.21.2",
+    "chokidar": "1.6.0",
+    "chromedriver": "2.22.1",
     "elasticdump": "2.1.1",
     "eslint": "1.10.3",
     "eslint-plugin-mocha": "1.1.0",
     "event-stream": "3.3.2",
     "expect.js": "0.3.1",
     "faker": "1.1.0",
-    "grunt": "0.4.5",
+    "grunt": "1.0.1",
+    "grunt-aws-s3": "0.14.5",
     "grunt-babel": "5.0.1",
     "grunt-cli": "0.1.13",
-    "grunt-contrib-clean": "0.6.0",
+    "grunt-contrib-clean": "1.0.0",
     "grunt-contrib-copy": "0.8.1",
-    "grunt-esvm": "3.2.1",
-    "grunt-karma": "0.12.0",
-    "grunt-run": "0.5.0",
-    "grunt-s3": "0.2.0-alpha.3",
+    "grunt-esvm": "3.2.4",
+    "grunt-karma": "2.0.0",
+    "grunt-run": "0.6.0",
     "grunt-simple-mocha": "0.4.0",
     "gruntify-eslint": "1.0.1",
     "handlebars": "4.0.5",
@@ -183,29 +183,31 @@
     "image-diff": "1.6.0",
     "intern": "3.2.3",
     "istanbul-instrumenter-loader": "0.1.3",
-    "karma": "0.13.9",
+    "karma": "1.2.0",
     "karma-chrome-launcher": "0.2.0",
     "karma-coverage": "0.5.1",
     "karma-firefox-launcher": "0.1.6",
     "karma-ie-launcher": "0.2.0",
     "karma-mocha": "0.2.0",
     "karma-safari-launcher": "0.1.1",
-    "license-checker": "3.1.0",
-    "load-grunt-config": "0.19.1",
+    "license-checker": "5.1.2",
+    "load-grunt-config": "0.19.2",
     "makelogs": "3.0.2",
-    "mocha": "2.3.0",
+    "marked-text-renderer": "0.1.0",
+    "mocha": "2.5.3",
     "ncp": "2.0.0",
     "nock": "2.10.0",
-    "npm": "2.15.8",
+    "npm": "3.10.3",
     "portscanner": "1.0.0",
     "simple-git": "1.37.0",
     "sinon": "1.17.2",
-    "source-map": "0.4.4",
-    "source-map-support": "0.4.0",
+    "source-map": "0.5.6",
+    "source-map-support": "0.2.10",
+    "supertest": "1.2.0",
     "supertest-as-promised": "2.0.2"
   },
   "engines": {
-    "node": "4.4.7",
-    "npm": "2.15.8"
+    "node": "6.4.0",
+    "npm": "3.10.3"
   }
 }
