@@ -1,6 +1,4 @@
 import Joi from 'joi';
-import fs from 'fs';
-import path from 'path';
 import { get } from 'lodash';
 import { randomBytes } from 'crypto';
 import os from 'os';
@@ -32,11 +30,11 @@ module.exports = () => Joi.object({
     exclusive: Joi.boolean().default(false)
   }).default(),
 
-  uuid: Joi.string().guid().default(),
 
   server: Joi.object({
+    uuid: Joi.string().guid().default(),
     name: Joi.string().default(os.hostname()),
-    host: Joi.string().hostname().default('0.0.0.0'),
+    host: Joi.string().hostname().default('localhost'),
     port: Joi.number().default(5601),
     maxPayloadBytes: Joi.number().default(1048576),
     autoListen: Joi.boolean().default(true),
