@@ -18,7 +18,7 @@ export default function (kbnServer, server, config) {
     handler: function (request, reply) {
       return reply({
         name: config.get('server.name'),
-        uuid: config.get('uuid'),
+        uuid: config.get('server.uuid'),
         status: kbnServer.status.toJSON(),
         metrics: kbnServer.metrics
       });
