@@ -10,11 +10,13 @@ import 'ui/timefilter';
 import 'ui/notify';
 import 'ui/private';
 import 'ui/promises';
+import 'ui/storage';
 import 'ui/directives/kbn_src';
 import 'ui/watch_multi';
+import './services';
 
-let chrome = {};
-let internals = _.defaults(
+const chrome = {};
+const internals = _.defaults(
   _.cloneDeep(metadata),
   {
     basePath: '',
@@ -28,10 +30,60 @@ let internals = _.defaults(
   }
 );
 
-$('<link>').attr({
-  href: require('ui/images/elk.ico'),
-  rel: 'shortcut icon'
-}).appendTo('head');
+// These favicons were generated from http://realfavicongenerator.net/.
+const favicons = [{
+  tag: '<link>',
+  attrs: {
+    href: require('ui/favicons/favicon.ico'),
+    rel: 'shortcut icon',
+  },
+}, {
+  tag: '<link>',
+  attrs: {
+    href: require('ui/favicons/apple-touch-icon.png'),
+    sizes: '180x180',
+    rel: 'apple-touch-icon',
+  },
+}, {
+  tag: '<link>',
+  attrs: {
+    href: require('ui/favicons/favicon-32x32.png'),
+    type: 'image/png',
+    sizes: '32x32',
+    rel: 'icon',
+  },
+}, {
+  tag: '<link>',
+  attrs: {
+    href: require('ui/favicons/favicon-16x16.png'),
+    type: 'image/png',
+    sizes: '16x16',
+    rel: 'icon',
+  },
+}, {
+  tag: '<link>',
+  attrs: {
+    href: require('ui/favicons/manifest.json'),
+    rel: 'manifest',
+  },
+}, {
+  tag: '<link>',
+  attrs: {
+    href: require('ui/favicons/safari-pinned-tab.svg'),
+    color: '#e8488b',
+    rel: 'mask-icon',
+  },
+}, {
+  tag: '<meta>',
+  attrs: {
+    name: 'theme-color',
+    content: '#e8488b',
+  },
+}];
+
+favicons.forEach(favicon => {
+  $(favicon.tag).attr(favicon.attrs).appendTo('head');
+});
 
 require('./api/apps')(chrome, internals);
 require('./api/xsrf')(chrome, internals);
