@@ -1,5 +1,6 @@
-import './app_switcher';
-import './app_switcher_link';
+
+import './global_nav';
+
 import kbnChromeProv from './kbn_chrome';
 import kbnChromeNavControlsProv from './append_nav_controls';
 import './kbn_loading_indicator';
