@@ -4,9 +4,9 @@ import _ from 'lodash';
 export default function () {
   let self = this;
   let scriptFields = {};
-  let fielddataFields = [];
+  let docvalueFields = [];
 
-  fielddataFields = _.pluck(self.fields.byType.date, 'name');
+  docvalueFields = _.map(_.reject(self.fields.byType.date, 'scripted'), 'name');
 
   _.each(self.getScriptedFields(), function (field) {
     scriptFields[field.name] = {
@@ -21,7 +21,7 @@ export default function () {
     storedFields: ['*'],
     _source: true,
     scriptFields: scriptFields,
-    fielddataFields: fielddataFields
+    docvalueFields: docvalueFields
   };
 
 };
