@@ -118,6 +118,15 @@ describe('Notifier', function () {
     it('includes stack', function () {
       expect(notify('error').stack).to.be.defined;
     });
+
+    it('has css class helper functions', function () {
+      expect(notify('error').getIconClass()).to.equal('fa fa-warning');
+      expect(notify('error').getButtonClass()).to.equal('btn-danger');
+      expect(notify('error').getAlertClassStack()).to.equal('toast-stack alert alert-danger');
+      expect(notify('error').getAlertClass()).to.equal('toast alert alert-danger');
+      expect(notify('error').getButtonGroupClass()).to.equal('toast-controls');
+      expect(notify('error').getToastMessageClass()).to.equal('toast-message');
+    });
   });
 
   describe('#warning', function () {
@@ -155,6 +164,15 @@ describe('Notifier', function () {
 
     it('does not include stack', function () {
       expect(notify('warning').stack).not.to.be.defined;
+    });
+
+    it('has css class helper functions', function () {
+      expect(notify('warning').getIconClass()).to.equal('fa fa-warning');
+      expect(notify('warning').getButtonClass()).to.equal('btn-warning');
+      expect(notify('warning').getAlertClassStack()).to.equal('toast-stack alert alert-warning');
+      expect(notify('warning').getAlertClass()).to.equal('toast alert alert-warning');
+      expect(notify('warning').getButtonGroupClass()).to.equal('toast-controls');
+      expect(notify('warning').getToastMessageClass()).to.equal('toast-message');
     });
   });
 
@@ -194,6 +212,15 @@ describe('Notifier', function () {
     it('does not include stack', function () {
       expect(notify('info').stack).not.to.be.defined;
     });
+
+    it('has css class helper functions', function () {
+      expect(notify('info').getIconClass()).to.equal('fa fa-info-circle');
+      expect(notify('info').getButtonClass()).to.equal('btn-info');
+      expect(notify('info').getAlertClassStack()).to.equal('toast-stack alert alert-info');
+      expect(notify('info').getAlertClass()).to.equal('toast alert alert-info');
+      expect(notify('info').getButtonGroupClass()).to.equal('toast-controls');
+      expect(notify('info').getToastMessageClass()).to.equal('toast-message');
+    });
   });
 
   describe('#custom', function () {
@@ -216,7 +243,7 @@ describe('Notifier', function () {
         customNotification = notifier.custom(customText, badParam);
       }
       expect(callCustomIncorrectly).to.throwException(function (e) {
-        expect(e.message).to.be('config param is required, and must be an object');
+        expect(e.message).to.be('Config param is required, and must be an object');
       });
 
     });
@@ -229,16 +256,16 @@ describe('Notifier', function () {
       // destroy the default custom notification, avoid duplicate handling
       customNotification.clear();
 
-      const explicitLifetimeParams = _.defaults({ lifetime: 20000 }, customParams);
-      customNotification = notifier.custom(customText, explicitLifetimeParams);
+      const overrideParams = _.defaults({ lifetime: 20000 }, customParams);
+      customNotification = notifier.custom(customText, overrideParams);
 
       expect(customNotification).to.have.property('type', 'info'); // default
-      expect(customNotification).to.have.property('title', explicitLifetimeParams.title); // passed in
-      expect(customNotification).to.have.property('lifetime', explicitLifetimeParams.lifetime); // passed in
+      expect(customNotification).to.have.property('title', overrideParams.title); // passed in thru customParams
+      expect(customNotification).to.have.property('lifetime', overrideParams.lifetime); // passed in thru overrideParams
 
-      expect(explicitLifetimeParams.type).to.be(undefined);
-      expect(explicitLifetimeParams.title).to.be.a('string');
-      expect(explicitLifetimeParams.lifetime).to.be.a('number');
+      expect(overrideParams.type).to.be(undefined);
+      expect(overrideParams.title).to.be.a('string');
+      expect(overrideParams.lifetime).to.be.a('number');
     });
 
     it('sets the content', function () {
@@ -249,6 +276,18 @@ describe('Notifier', function () {
     it('uses custom actions', function () {
       expect(customNotification).to.have.property('customActions');
       expect(customNotification.customActions).to.have.length(customParams.actions.length);
+    });
+
+    it('custom actions have getButtonClass method', function () {
+      customNotification.customActions.forEach((action, idx) => {
+        expect(action).to.have.property('getButtonClass');
+        expect(action.getButtonClass).to.be.a('function');
+        if (idx === 0) {
+          expect(action.getButtonClass()).to.be('btn-primary btn-info');
+        } else {
+          expect(action.getButtonClass()).to.be('btn-default btn-info');
+        }
+      });
     });
 
     it('gives a default action if none are provided', function () {
@@ -357,6 +396,15 @@ describe('Notifier', function () {
     it('does not include stack', function () {
       expect(notify('banner').stack).not.to.be.defined;
     });
+
+    it('has css class helper functions', function () {
+      expect(notify('banner').getIconClass()).to.equal('');
+      expect(notify('banner').getButtonClass()).to.equal('btn-banner');
+      expect(notify('banner').getAlertClassStack()).to.equal('toast-stack alert alert-banner');
+      expect(notify('banner').getAlertClass()).to.equal('alert alert-banner');
+      expect(notify('banner').getButtonGroupClass()).to.equal('toast-controls-banner');
+      expect(notify('banner').getToastMessageClass()).to.equal('toast-message-banner');
+    });
   });
 
   function notify(fnName) {
@@ -382,4 +430,180 @@ describe('Notifier', function () {
       });
     });
   }
+});
+
+describe('Directive Notification', function () {
+  let notifier;
+  let compile;
+  let scope;
+
+  const directiveParam = {
+    template: '<h1>Hello world {{ unit.message }}</h1>',
+    controllerAs: 'unit',
+    controller() {
+      this.message = '🎉';
+    }
+  };
+  const customParams = {
+    title: 'fooTitle',
+    actions:[{
+      text: 'Cancel',
+      callback: sinon.spy()
+    }, {
+      text: 'OK',
+      callback: sinon.spy()
+    }]
+  };
+  let directiveNotification;
+
+  beforeEach(() => {
+    ngMock.module('kibana');
+    ngMock.inject(function ($rootScope, $compile) {
+      scope = $rootScope.$new();
+      compile = $compile;
+      compile;
+      scope;
+    });
+
+    while (Notifier.prototype._notifs.pop()); // clear global notifications
+
+    notifier = new Notifier({ location: 'directiveFoo' });
+    directiveNotification = notifier.directive(directiveParam, customParams);
+  });
+
+  afterEach(() => {
+    directiveNotification.clear();
+    scope.$destroy();
+  });
+
+  describe('returns a renderable notification', () => {
+    let element;
+
+    beforeEach(() => {
+      scope.notif = notifier.directive(directiveParam, customParams);
+      const template = `
+        <render-directive
+          definition="notif.directive"
+          notif="notif"
+        ></render-directive>`;
+      element = compile(template)(scope);
+      scope.$apply();
+    });
+
+    it('that renders with the provided template', () => {
+      expect(element.find('h1').text()).to.contain('Hello world');
+    });
+
+    it('that renders with the provided controller', () => {
+      expect(element.text()).to.contain('🎉');
+    });
+  });
+
+  it('throws if first param is not an object', () => {
+    // destroy the default custom notification, avoid duplicate handling
+    directiveNotification.clear();
+
+    function callDirectiveIncorrectly() {
+      const badDirectiveParam = null;
+      directiveNotification = notifier.directive(badDirectiveParam, {});
+    }
+    expect(callDirectiveIncorrectly).to.throwException(function (e) {
+      expect(e.message).to.be('Directive param is required, and must be an object');
+    });
+  });
+
+  it('throws if second param is not an object', () => {
+    // destroy the default custom notification, avoid duplicate handling
+    directiveNotification.clear();
+
+    function callDirectiveIncorrectly() {
+      const badConfigParam = null;
+      directiveNotification = notifier.directive(directiveParam, badConfigParam);
+    }
+    expect(callDirectiveIncorrectly).to.throwException(function (e) {
+      expect(e.message).to.be('Config param is required, and must be an object');
+    });
+  });
+
+  it('throws if directive param has scope definition instead of allow the helper to do its work', () => {
+    // destroy the default custom notification, avoid duplicate handling
+    directiveNotification.clear();
+
+    function callDirectiveIncorrectly() {
+      const badDirectiveParam = {
+        scope: {
+          garbage: '='
+        }
+      };
+      directiveNotification = notifier.directive(badDirectiveParam, customParams);
+    }
+    expect(callDirectiveIncorrectly).to.throwException(function (e) {
+      expect(e.message).to.be('Directive should not have a scope definition. Notifier has an internal implementation.');
+    });
+  });
+
+  it('throws if directive param has link function instead of allow the helper to do its work', () => {
+    // destroy the default custom notification, avoid duplicate handling
+    directiveNotification.clear();
+
+    function callDirectiveIncorrectly() {
+      const badDirectiveParam = {
+        link: ($scope) => {
+          /*eslint-disable no-console*/
+          console.log($scope.nothing);
+          /*eslint-enable*/
+        }
+      };
+      directiveNotification = notifier.directive(badDirectiveParam, customParams);
+    }
+    expect(callDirectiveIncorrectly).to.throwException(function (e) {
+      expect(e.message).to.be('Directive should not have a link function. Notifier has an internal link function helper.');
+    });
+  });
+
+  it('has a directive function to make notifications with template and scope', () => {
+    expect(notifier.directive).to.be.a('function');
+  });
+
+  it('sets the scope property and link function', () => {
+    expect(directiveNotification).to.have.property('directive');
+    expect(directiveNotification.directive).to.be.an('object');
+
+    expect(directiveNotification.directive).to.have.property('scope');
+    expect(directiveNotification.directive.scope).to.be.an('object');
+
+    expect(directiveNotification.directive).to.have.property('link');
+    expect(directiveNotification.directive.link).to.be.an('function');
+  });
+
+  /* below copied from custom notification tests */
+  it('uses custom actions', () => {
+    expect(directiveNotification).to.have.property('customActions');
+    expect(directiveNotification.customActions).to.have.length(customParams.actions.length);
+  });
+
+  it('gives a default action if none are provided', () => {
+    // destroy the default custom notification, avoid duplicate handling
+    directiveNotification.clear();
+
+    const noActionParams = _.defaults({ actions: [] }, customParams);
+    directiveNotification = notifier.directive(directiveParam, noActionParams);
+    expect(directiveNotification).to.have.property('actions');
+    expect(directiveNotification.actions).to.have.length(1);
+  });
+
+  it('defaults type and lifetime for "info" config', () => {
+    expect(directiveNotification.type).to.be('info');
+    expect(directiveNotification.lifetime).to.be(5000);
+  });
+
+  it('should wrap the callback functions in a close function', () => {
+    directiveNotification.customActions.forEach((action, idx) => {
+      expect(action.callback).not.to.equal(customParams.actions[idx]);
+      action.callback();
+    });
+    customParams.actions.forEach(action => {
+      expect(action.callback.called).to.true;
+    });
+  });
 });
