@@ -1,8 +1,10 @@
 import _ from 'lodash';
+import angular from 'angular';
 import $ from 'jquery';
 import metadata from 'ui/metadata';
 import formatMsg from 'ui/notify/lib/_format_msg';
 import fatalSplashScreen from 'ui/notify/partials/fatal_splash_screen.html';
+import 'ui/render_directive';
 /* eslint no-console: 0 */
 
 let notifs = [];
@@ -83,6 +85,27 @@ function restartNotifTimer(notif, cb) {
   startNotifTimer(notif, cb);
 }
 
+const typeToButtonClassMap = {
+  danger: 'btn-danger', // NOTE: `error` type is internally named as `danger`
+  warning: 'btn-warning',
+  info: 'btn-info',
+  banner: 'btn-banner'
+};
+const buttonHierarchyClass = (index) => {
+  if (index === 0) {
+    // first action: primary className
+    return 'btn-primary';
+  }
+  // subsequent actions: secondary/default className
+  return 'btn-default';
+};
+const typeToAlertClassMap = {
+  danger: `alert-danger`,
+  warning: `alert-warning`,
+  info: `alert-info`,
+  banner: `alert-banner`,
+};
+
 function add(notif, cb) {
   _.set(notif, 'info.version', version);
   _.set(notif, 'info.buildNum', buildNum);
@@ -95,10 +118,14 @@ function add(notif, cb) {
     });
   } else if (notif.customActions) {
     // wrap all of the custom functions in a close
-    notif.customActions = notif.customActions.map(action => {
+    notif.customActions = notif.customActions.map((action, index) => {
       return {
         key: action.text,
-        callback: closeNotif(notif, action.callback, action.text)
+        callback: closeNotif(notif, action.callback, action.text),
+        getButtonClass() {
+          const buttonTypeClass = typeToButtonClassMap[notif.type];
+          return `${buttonHierarchyClass(index)} ${buttonTypeClass}`;
+        }
       };
     });
   }
@@ -109,9 +136,24 @@ function add(notif, cb) {
     return notif.timerId ? true : false;
   };
 
-  let dup = _.find(notifs, function (item) {
-    return item.content === notif.content && item.lifetime === notif.lifetime;
-  });
+  // decorate the notification with helper functions for the template
+  notif.getButtonClass = () => typeToButtonClassMap[notif.type];
+  notif.getAlertClassStack = () => `toast-stack alert ${typeToAlertClassMap[notif.type]}`;
+  notif.getIconClass = () => (notif.type === 'banner') ?  '' : `fa fa-${notif.icon}`;
+  notif.getToastMessageClass = ()  => (notif.type === 'banner') ?  'toast-message-banner' : 'toast-message';
+  notif.getAlertClass = () => (notif.type === 'banner') ?
+    `alert ${typeToAlertClassMap[notif.type]}` : // not including `.toast` class leaves out the flex properties for banner
+    `toast alert ${typeToAlertClassMap[notif.type]}`;
+  notif.getButtonGroupClass = () => (notif.type === 'banner') ?
+    'toast-controls-banner' :
+    'toast-controls';
+
+  let dup = null;
+  if (notif.content) {
+    dup = _.find(notifs, function (item) {
+      return item.content === notif.content && item.lifetime === notif.lifetime;
+    });
+  }
 
   if (dup) {
     dup.count += 1;
@@ -400,28 +442,13 @@ Notifier.prototype.banner = function (msg, cb) {
 };
 
 /**
- * Display a custom message
- * @param  {String} msg - required
- * @param  {Object} config - required
- * @param  {Function} cb - optional
- *
- * config = {
- *   title: 'Some Title here',
- *   type: 'info',
- *   actions: [{
- *     text: 'next',
- *     callback: function() { next(); }
- *   }, {
- *     text: 'prev',
- *     callback: function() { prev(); }
- *   }]
- * }
+ * Helper for common behavior in custom and directive types
  */
-Notifier.prototype.custom = function (msg, config, cb) {
+function getDecoratedCustomConfig(config) {
   // There is no helper condition that will allow for 2 parameters, as the
   // other methods have. So check that config is an object
   if (!_.isPlainObject(config)) {
-    throw new Error('config param is required, and must be an object');
+    throw new Error('Config param is required, and must be an object');
   }
 
   // workaround to allow callers to send `config.type` as `error` instead of
@@ -444,22 +471,103 @@ Notifier.prototype.custom = function (msg, config, cb) {
     }
   };
 
-  const mergedConfig = _.assign({
+  const customConfig = _.assign({
     type: 'info',
     title: 'Notification',
-    content: formatMsg(msg, this.from),
     lifetime: getLifetime(config.type)
   }, config);
 
-  const hasActions = _.get(mergedConfig, 'actions.length');
+  const hasActions = _.get(customConfig, 'actions.length');
   if (hasActions) {
-    mergedConfig.customActions = mergedConfig.actions;
-    delete mergedConfig.actions;
+    customConfig.customActions = customConfig.actions;
+    delete customConfig.actions;
   } else {
-    mergedConfig.actions = ['accept'];
+    customConfig.actions = ['accept'];
   }
 
-  return add(mergedConfig, cb);
+  return customConfig;
+}
+
+/**
+ * Display a custom message
+ * @param  {String} msg - required
+ * @param  {Object} config - required
+ * @param  {Function} cb - optional
+ *
+ * config = {
+ *   title: 'Some Title here',
+ *   type: 'info',
+ *   actions: [{
+ *     text: 'next',
+ *     callback: function() { next(); }
+ *   }, {
+ *     text: 'prev',
+ *     callback: function() { prev(); }
+ *   }]
+ * }
+ */
+Notifier.prototype.custom = function (msg, config, cb) {
+  const customConfig = getDecoratedCustomConfig(config);
+  customConfig.content = formatMsg(msg, this.from);
+  return add(customConfig, cb);
+};
+
+/**
+ * Display a scope-bound directive using template rendering in the message area
+ * @param  {Object} directive - required
+ * @param  {Object} config - required
+ * @param  {Function} cb - optional
+ *
+ * directive = {
+ *  template: `<p>Hello World! <a ng-click="example.clickHandler()">Click me</a>.`,
+ *  controllerAs: 'example',
+ *  controller() {
+ *    this.clickHandler = () {
+ *      // do something
+ *    };
+ *  }
+ * }
+ *
+ * config = {
+ *   title: 'Some Title here',
+ *   type: 'info',
+ *   actions: [{
+ *     text: 'next',
+ *     callback: function() { next(); }
+ *   }, {
+ *     text: 'prev',
+ *     callback: function() { prev(); }
+ *   }]
+ * }
+ */
+Notifier.prototype.directive = function (directive, config, cb) {
+  if (!_.isPlainObject(directive)) {
+    throw new Error('Directive param is required, and must be an object');
+  }
+  if (!Notifier.$compile) {
+    throw new Error('Unable to use the directive notification until Angular has initialized.');
+  }
+  if (directive.scope) {
+    throw new Error('Directive should not have a scope definition. Notifier has an internal implementation.');
+  }
+  if (directive.link) {
+    throw new Error('Directive should not have a link function. Notifier has an internal link function helper.');
+  }
+
+  // make a local copy of the directive param (helps unit tests)
+  const localDirective = _.clone(directive, true);
+
+  localDirective.scope = { notif: '=' };
+  localDirective.link = function link($scope, $el) {
+    const $template = angular.element($scope.notif.directive.template);
+    const postLinkFunction = Notifier.$compile($template);
+    $el.html($template);
+    postLinkFunction($scope);
+  };
+
+  const customConfig = getDecoratedCustomConfig(config);
+  customConfig.directive = localDirective;
+  return add(customConfig, cb);
 };
 
 Notifier.prototype.describeError = formatMsg.describeError;
@@ -514,6 +622,7 @@ function createGroupLogger(type, opts) {
 
     if (consoleGroups) {
       if (status) {
+        console.log(status);
         console.groupEnd();
       } else {
         if (opts.open) {
