@@ -3,12 +3,12 @@ import 'ui/filters/uriescape';
 import 'ui/filters/rison';
 import uiModules from 'ui/modules';
 import rison from 'rison-node';
-
+import AppStateProvider from 'ui/state_management/app_state';
 
 uiModules.get('kibana/url')
 .service('kbnUrl', function (Private) { return Private(KbnUrlProvider); });
 
-function KbnUrlProvider($route, $location, $rootScope, globalState, $parse, getAppState) {
+function KbnUrlProvider($injector, $location, $rootScope, $parse, Private) {
   let self = this;
 
   /**
@@ -162,21 +162,25 @@ function KbnUrlProvider($route, $location, $rootScope, globalState, $parse, getA
       search: $location.search()
     };
 
-    if (self._shouldAutoReload(next, prev)) {
-      let appState = getAppState();
-      if (appState) appState.destroy();
+    if ($injector.has('$route')) {
+      const $route = $injector.get('$route');
 
-      reloading = $rootScope.$on('$locationChangeSuccess', function () {
-        // call the "unlisten" function returned by $on
-        reloading();
-        reloading = false;
+      if (self._shouldAutoReload(next, prev, $route)) {
+        const appState = Private(AppStateProvider).getAppState();
+        if (appState) appState.destroy();
 
-        $route.reload();
-      });
+        reloading = $rootScope.$on('$locationChangeSuccess', function () {
+          // call the "unlisten" function returned by $on
+          reloading();
+          reloading = false;
+
+          $route.reload();
+        });
+      }
     }
   };
 
-  self._shouldAutoReload = function (next, prev) {
+  self._shouldAutoReload = function (next, prev, $route) {
     if (reloading) return false;
 
     let route = $route.current && $route.current.$$route;
