@@ -9,18 +9,22 @@ describe('slugifyId()', function () {
     ['test?test', 'test-questionmark-test'],
     ['test=test', 'test-equal-test'],
     ['test&test', 'test-ampersand-test'],
+    ['test%test', 'test-percent-test'],
     ['test / test', 'test-slash-test'],
     ['test ? test', 'test-questionmark-test'],
     ['test = test', 'test-equal-test'],
     ['test & test', 'test-ampersand-test'],
+    ['test % test', 'test-percent-test'],
     ['test / ^test', 'test-slash-^test'],
     ['test ?  test', 'test-questionmark-test'],
     ['test =  test', 'test-equal-test'],
     ['test &  test', 'test-ampersand-test'],
+    ['test %  test', 'test-percent-test'],
     ['test/test/test', 'test-slash-test-slash-test'],
     ['test?test?test', 'test-questionmark-test-questionmark-test'],
     ['test&test&test', 'test-ampersand-test-ampersand-test'],
-    ['test=test=test', 'test-equal-test-equal-test']
+    ['test=test=test', 'test-equal-test-equal-test'],
+    ['test%test%test', 'test-percent-test-percent-test']
   ];
 
   _.each(fixtures, function (fixture) {
