@@ -2,62 +2,44 @@ module.exports = function (grunt) {
   return {
     options: {
       licenses: [
-        'MIT',
-        'MIT*',
-        'MIT License',
-        'MIT/X11',
-        'OFL-1.1 AND MIT',
-        'new BSD, and MIT',
-        '(MIT OR Apache-2.0)',
-        'BSD',
-        'BSD*',
-        'BSD New',
-        'BSD-like',
-        'BSD-2-Clause',
         '(BSD-2-Clause OR MIT OR Apache-2.0)',
+        '(BSD-2-Clause OR MIT)',
+        '(MIT AND CC-BY-3.0)',
+        '(MIT OR Apache-2.0)',
+        'AFLv2.1',
+        'Apache',
+        'Apache 2.0',
+        'Apache License, v2.0',
+        'Apache*',
+        'Apache, Version 2.0',
+        'Apache-2.0',
+        'BSD',
+        'BSD New',
+        'BSD*',
+        'BSD-2-Clause',
         'BSD-3-Clause',
+        'BSD-3-Clause AND MIT',
+        'BSD-like',
         'CC-BY',
         'CC-BY-4.0',
-        'Apache',
-        'Apache*',
-        'Apache v2',
-        'Apache 2.0',
-        'Apache2',
-        'Apache-2.0',
-        'Apache, Version 2.0',
-        'Apache License, v2.0',
         'ISC',
-        'WTFPL',
-        'Public-Domain',
+        'MIT',
+        'MIT*',
+        'MIT/X11',
+        'OFL-1.1 AND MIT',
         'Public domain',
-        'UNLICENSE'
+        'WTFPL',
+        'new BSD, and MIT'
       ],
       overrides: {
-        'amdefine@1.0.0': ['BSD-3-Clause', 'MIT'],
-        '@spalger/angular-bootstrap@0.12.1': ['MIT'],
-        'angular-ui-ace@0.2.3': ['MIT'],
         'assert-plus@0.1.5': ['MIT'],
         'buffers@0.1.1': ['MIT/X11'],
-        'color-name@1.0.0': ['UNLICENSE'],
-        'commander@2.2.0': ['MIT'],
-        'css-color-names@0.0.1': ['MIT'],
         'css-parse@1.0.4': ['MIT'],
         'css-stringify@1.0.5': ['MIT'],
         'css@1.0.8': ['MIT'],
-        'cycle@1.0.3': ['Public-Domain'],
         'delegate@3.0.1': ['MIT'],
-        'FileSaver@1.1.0': ['MIT'],
         'flatten@0.0.1': ['MIT'],
         'indexof@0.0.1': ['MIT'],
-        'inherits@1.0.0': ['ISC'],
-        'is-buffer@1.1.2': ['MIT'],
-        'jsonpointer@1.1.0': ['MIT'],
-        'lazy-cache@1.0.3': ['MIT'],
-        'leaflet@0.7.2': ['BSD-2-Clause'],
-        'Nonsense@0.1.2': ['Public-Domain'],
-        'pkginfo@0.2.3': ['MIT'],
-        'readable-stream@2.0.5': ['MIT'],
-        'repeat-string@1.5.2': ['MIT'],
         'ripemd160@0.2.0': ['MIT'],
         'select@1.0.6': ['MIT'],
         'uglify-js@2.2.5': ['BSD'],
