@@ -68,12 +68,39 @@ export default (function () {
           );
         } else {
           configId = response.hits.hits[0]._id;
-          Log.debug('config._id =' + configId);
+          Log.debug('config._id = ' + configId);
           return configId;
         }
       });
     },
 
+    /*
+    ** Gets defaultIndex from the config doc.
+    */
+    getDefaultIndex: function () {
+      var defaultIndex;
+
+      return this.client.search({
+        index: '.kibana',
+        type: 'config'
+      })
+      .then(function (response) {
+        if (response.errors) {
+          throw new Error(
+            'get config failed\n' +
+              response.items
+              .map(i => i[Object.keys(i)[0]].error)
+              .filter(Boolean)
+              .map(err => '  ' + JSON.stringify(err))
+              .join('\n')
+          );
+        } else {
+          defaultIndex = response.hits.hits[0]._source.defaultIndex;
+          Log.debug('config.defaultIndex = ' + defaultIndex);
+          return defaultIndex;
+        }
+      });
+    },
 
     /**
     * Add fields to the config doc (like setting timezone and defaultIndex)
